secure-vault-api/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
firebase-admin>=6.0.0
python-multipart>=0.0.6
pydantic>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
requests>=2.28.0
